# example-microservice/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Delivery Specification
  description: This is a microservice for managing delivery in the YumYumNow company.
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /status/{orderId}:
    get:
      tags:
        - Status
      summary: Get the order status
      operationId: getStatus
      description: Get the status of an order
      parameters:
        - in: path
          name: orderId
          schema:
            type: integer
            format: int64
          required: true
          description: Id of the order to get status from
        - in: query
          name: authorization
          description: The userId to check if they have the rights to make a request
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved the order
          content:
            application/json:
              schema:
                type: string
                enum:
                  - pending
                  - accepted
                  - rejected
                  - preparing
                  - given-to-courier
                  - in-transit
                  - delivered
        '400':
          description: Invalid arguments
        '404':
          description: Order not found
        '403':
          description: Unauthorized
  /status/{orderId}/accepted:
    put:
      tags:
        - Status
      summary: Change the order status from pending to accepted
      operationId: updateToAccepted
      description: Update the order status to accepted
      parameters:
        - in: path
          name: orderId
          schema:
            type: integer
            format: int64
          required: true
          description: Id of the order to change status
        - in: query
          name: authorization
          description: The userId to check if they have the rights to make a request
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful update of status to accepted
        '400':
          description: Invalid arguments
        '404':
          description: Order not found
        '403':
          description: Unauthorized
  /status/{orderId}/rejected:
    put:
      summary: Set status to “rejected”
      operationId: updateToRejected
      tags:
        - Status
      description: Sends an exception, goes from “pending” to [said by vendor] “rejected”
      parameters:
        - in: path
          name: orderId
          description: Id of the order to change status
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: authorization
          description: the UserId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful update of status to rejected
        '400':
          description: Invalid arguments
        '403':
          description: Unauthorized
        '404':
          description: Order not found
  /status/{orderId}/preparing:
    put:
      tags:
        - Status
      summary: Change the order status from accepted to preparing
      operationId: updateToPreparing
      description: Update the order status to preparing
      parameters:
        - in: path
          name: orderId
          schema:
            type: integer
            format: int64
          required: true
          description: Id of order to change status
        - in: query
          name: authorization
          description: The userId to check if they have the rights to make a request
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/PreparingRequestBody'
      responses:
        '200':
          description: Successful update of status to preparing
        '400':
          description: Invalid arguments
        '404':
          description: Order not found
        '403':
          description: Unauthorized
  /status/{orderId}/giventocourier:
    put:
      summary: Set status to “given to courier”
      operationId: updateToGivenToCourier
      tags:
        - Status
      description: Goes from “preparing” to [said by vendor] “given to courier” and set the courierID
      requestBody:
        $ref: '#/components/requestBodies/GivenToCourierRequestBody'
      parameters:
        - in: path
          name: orderId
          description: Id of the order to change status
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: authorization
          description: the UserId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful update of status to "given to courier"
        '400':
          description: Invalid arguments
        '403':
          description: Unauthorized
        '404':
          description: Order not found
  /status/{orderId}/intransit:
    put:
      tags:
        - Status
      summary: Change the order status from accepted to in-transit
      description: Update the order status to in-transit
      operationId: updateToInTransit
      parameters:
        - in: path
          name: orderId
          schema:
            type: integer
            format: int64
          required: true
          description: Id of order to change status
        - in: query
          name: authorization
          description: The userId to check if they have the rights to make a request
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful update of status to in-transit
        '400':
          description: Invalid arguments
        '404':
          description: Order not found
        '403':
          description: Unauthorized
...
components:
...
  requestBodies:
    PreparingRequestBody:
      description: Request body for status change from accepted to preparing
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              prepTime:
                type: string
                format: time
                description: the time it takes to prepare the order in format "HH:MM:SS"
                example: '01:30:00'
              expectedDeliveryTime:
                type: string
                format: date-time
                description: the time the order will be planned to be delivered
    GivenToCourierRequestBody:
      description: Request body for status change from preparing to given to courier
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              courierId:
                type: integer
                format: int64
                example: 88
...